requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
blinker==1.9.0
    # via flask
certifi==2026.4.22
    # via requests
chardet==5.2.0
    # via -r requirements.in
charset-normalizer==3.4.7
    # via requests
click==8.4.0
    # via flask
colorama==0.4.6
    # via click
et-xmlfile==2.0.0
    # via openpyxl
flask==3.1.3
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
idna==3.15
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
mysql-connector-python==9.7.0
    # via -r requirements.in
openpyxl==3.1.5
    # via -r requirements.in
packaging==26.2
    # via gunicorn
psycopg2-binary==2.9.12
    # via -r requirements.in
python-dotenv==1.2.2
    # via -r requirements.in
requests==2.34.2
    # via -r requirements.in
urllib3==2.5.0
    # via
    #   -r requirements.in
    #   requests
werkzeug==3.1.8
    # via flask
